+++ src/presenter/extra-film-list.js
@@ -1,8 +1,8 @@
-import abstractfilmList from './abstract-film-list';
+import AbstractFilmList from './abstract-film-list';
 import ExtraFilmsBlock from '../view/films/extra-films-block.js';
 
 
-export default class ExtraFilmList extends abstractfilmList {
+export default class ExtraFilmList extends AbstractFilmList {
   constructor(container, name) {
     super(container);
