+++ src/view/films/no-films.js
@@ -1,14 +1,8 @@
 import Abstract from '../abstract.js';
 
-const EmptyResultMessage = {
-  ALL: 'There are no movies in our database',
-  WATCH_LIST: 'There are no movies to watch now',
-  HISTORY: 'There are no watched movies now',
-  FAVORITE: 'There are no favorite movies now',
-};
 
-const createNoFilm = (filter) => `
-<h2 class="films-list__title">${EmptyResultMessage[filter]}</h2>
+const createNoFilm = (message) => `
+<h2 class="films-list__title">${message}</h2>
 `;
 
 export default class NoFilms extends Abstract{
